Refactor AnalysisReport.
