+++ opossum_core/src/reporting/mod.rs
@@ -2,4 +2,3 @@
 pub mod analysis_report;
 pub mod html_report;
 pub mod node_report;
-pub mod report_helper;
